+++ src/query/storages/fuse/fuse/src/operations/mutation/mod.rs
@@ -16,7 +16,7 @@ pub mod abort_operation;
 pub mod base_mutator;
 pub mod block_filter;
 mod compact;
-mod deletion_mutation;
+mod deletion;
 pub mod deletion_mutator;
 pub mod recluster_mutator;
